Move element specific traits to sub modules
